refactor(web-serial-rxjs): extract port teardown helpers
Move teardownPump and closePortSafely out of the session lifecycle into a
dedicated port-teardown module so the error reporter can depend on them
without forming a construction cycle with the lifecycle.
